Add navigator
